Refactor statistics out of calendar
